clean up server.rb
